Allow local resources to use in app links
